Adds game type functionality
Introduces a game type field to differentiate between local and online games.

This change adds a "gameType" field to the GameModel and database schema. It allows for distinguishing between local and online game instances, which will be used for future feature implementations. The default game type is set to "local".