Adds game type selection to lobby
Allows users to filter games by type (local or online) in the lobby.
This enhances the user experience by enabling players to more easily find the type of game they are looking for.

Also fixes a bug that would cause the game to start before both players joined a local game.